+++ R/check.R
@@ -6,6 +6,7 @@ check = function(server) {
   cli::cli_h2("Checking Server")
   status_codes = check_status_code(server)
   headers = check_server_headers(server)
-  list(status_codes = status_codes,
-       headers = headers)
+  rtn = list(status_codes = status_codes,
+             headers = headers)
+  invisible(rtn)
 }